requirements.txt
FastAPI
uvicorn
python-dotenv
psycopg2-binary